RSA inital commit
This is the initial commit to the group project. Right now the embedded main driver code (RSA_REF.js) is setup to work with the demo RSA.html file in case we do not have something to demo come the second presentation. The driver code will be adapted to the main page once it has been finalized. Let me know if something is not working