Directory Revamp and Additions
Made a lib folder and got rid of lib.js. I put all the encryption stuff in there and a README which has a line of HTML showing how to reference the lib folder contents within the HTML. made AES(moved Jairo's files in there) and RSA folders within the lib folder. put bootstrap.js and jquery.js in the folder. Made some changes to the HTML and added jquery.js and bootstrap.js to the HTML.